python+reevit/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
reevit>=0.1.0
python-dotenv>=1.0.0
